WA-232 Refactor safe tests improving its speed
