+++ src/test/safe.dom.transactions.test.js
